# api/swagger.yaml
openapi: "3.0.2"
info:
  title: SAWM API
  version: "1.0"
servers:
  - url: http://localhost:3080
paths:
  /User/{token}:
    get:
      tags:
        - User
      summary: Get user information
      parameters:
        - name: token
          in: path
          description: Spotify access token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User information
          content:
            application/json:
              schema:
                  type: object
                  properties:
                    display_name:
                      type: string
                    external_urls:
                      type: object
                      properties:
                        spotify:
                          type: string
                    followers:
                      type: object
                      properties:
                        href:
                          type: string
                        total:
                          type: integer
                    href:
                      type: string
                    id:
                      type: string
                    images:
                      type: array
                      items:
                        type: object
                        properties:
                          url:
                            type: string
                          height:
                            type: integer
                          width:
                            type: integer
                    type:
                      type: string
                    uri:
                      type: string

        "401":
            description: unauthorized token
  /UserPlaylists/{token}:
    get:
      tags:
        - User
      summary: Get user playlists
      parameters:
        - name: token
          in: path
          description: Spotify access token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User playlists
          content:
            application/json:
              schema:
                type: object
                properties:
                  href:
                    type: string
                  limit:
                    type: integer
                  next:
                    type: string
                  offset:
                    type: integer
                  previous:
                    type: string
                  total:
                    type: integer
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        collaborative:
                          type: boolean
                        description:
                          type: string
                        external_urls:
                          type: object
                          properties:
                            spotify:
                              type: string
                        href:
                          type: string
                        id:
                          type: string
                        images:
                          type: array
                          items:
                            type: object
                            properties:
                              url:
                                type: string
                              height:
                                type: integer
                              width:
                                type: integer
                        name:
                          type: string
                        owner:
                          type: object
                          properties:
                            external_urls:
                              type: object
                              properties:
                                spotify:
                                  type: string
                            followers:
                              type: object
                              properties:
                                href:
                                  type: string
                                total:
                                  type: integer
                            href:
                              type: string
                            id:
                              type: string
                            type:
                              type: string
                            uri:
                              type: string
                            display_name:
                              type: string
                        public:
                          type: boolean
                        snapshot_id:
                          type: string
                        tracks:
                          type: object
                          properties:
                            href:
                              type: string
                            total:
                              type: integer
                        type:
                          type: string
                        uri:
                          type: string
        "401":
            description: unauthorized token
  /UserTopTracks/{time}/{token}:
    get:
      tags:
        - User
      summary: Get user top tracks
      parameters:
        - name: time
          in: path
          description: Time range for top tracks
          required: true
          schema:
            type: string
        - name: token
          in: path
          description: Spotify access token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User top tracks
          content:
            application/json:
              schema:
                type: object
                properties:
                  href:
                    type: string
                  limit:
                    type: integer
                  next:
                    type: string
                  offset:
                    type: integer
                  previous:
                    type: string
                  total:
                    type: integer
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        external_urls:
                          type: object
                          properties:
                            spotify:
                              type: string
                        followers:
                          type: object
                          properties:
                            href:
                              type: string
                            total:
                              type: integer
                        genres:
                          type: array
                          items:
                            type: string
                        href:
                          type: string
                        id:
                          type: string
                        images:
                          type: array
                          items:
                            type: object
                            properties:
                              url:
                                type: string
                              height:
                                type: integer
                              width:
                                type: integer
                        name:
                          type: string
                        popularity:
                          type: integer
                        type:
                          type: string
                        uri:
                          type: string
        "401":
            description: unauthorized token
  /UserTopArtists/{time}/{token}:
    get:
      tags:
        - User
      summary: Get user top artists
      parameters:
        - name: token
          in: path
          description: Spotify access token
          required: true
          schema:
            type: string
        - name: time
          in: query
          description: Time range for top artists
          required: false
          schema:
            type: string
      responses:
        "200":
          description: User top artists
          content:
            application/json:
              schema:
                type: object
                properties:
                  href:
                    type: string
                  limit:
                    type: integer
                  next:
                    type: string
                  offset:
                    type: integer
                  previous:
                    type: string
                  total:
                    type: integer
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        external_urls:
                          type: object
                          properties:
                            spotify:
                              type: string
                        followers:
                          type: object
                          properties:
                            href:
                              type: string
                            total:
                              type: integer
                        genres:
                          type: array
                          items:
                            type: string
                        href:
                          type: string
                        id:
                          type: string
                        images:
                          type: array
                          items:
                            type: object
                            properties:
                              url:
                                type: string
                              height:
                                type: integer
                              width:
                                type: integer
                        name:
                          type: string
                        popularity:
                          type: integer
                        type:
                          type: string
                        uri:
                          type: string
        "401":
            description: unauthorized token
  /UserAlbums/{token}:
    get:
      tags:
        - User
      summary: Get user albums
      parameters:
        - name: token
          in: path
          description: Spotify access token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User albums
          content:
            application/json:
              schema:
                type: object
                properties:
                  href:
                    type: string
                  limit:
                    type: integer
                  next:
                    type: string
                  offset:
                    type: integer
                  previous:
                    type: string
                  total:
                    type: integer
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        added_at:
                          type: string
                        album:
                          type: object
                          properties:
                            album_type:
                              type: string
                            total_tracks:
                              type: integer
                            available_markets:
                              type: array
                              items:
                                type: string
                            external_urls:
                              type: object
                              properties:
                                spotify:
                                  type: string
                            href:
                              type: string
                            id:
                              type: string
                            images:
                              type: array
                              items:
                                type: object
                                properties:
                                  url:
                                    type: string
                                  height:
                                    type: integer
                                  width:
                                    type: integer
                            name:
                              type: string
                            release_date:
                              type: string
                            release_date_precision:
                              type: string
                            restrictions:
                              type: object
                              properties:
                                reason:
                                  type: string
                            type:
                              type: string
                            uri:
                              type: string
                            copyrights:
                              type: array
                              items:
                                type: object
                                properties:
                                  text:
                                    type: string
                                  type:
                                    type: string
                            external_ids:
                              type: object
                              properties:
                                isrc:
                                  type: string
                                ean:
                                  type: string
                                upc:
                                  type: string
                            genres:
                              type: array
                              items:
                                type: string
                            label:
                              type: string
                            popularity:
                              type: integer
                            artists:
                              type: array
                              items:
                                type: object
                                properties:
                                  external_urls:
                                    type: object
                                    properties:
                                      spotify:
                                        type: string
                                  followers:
                                    type: object
                                    properties:
                                      href:
                                        type: string
                                      total:
                                        type: integer
                                  genres:
                                    type: array
                                    items:
                                      type: string
                                  href:
                                    type: string
                                  id:
                                    type: string
                                  images:
                                    type: array
                                    items:
                                      type: object
                                      properties:
                                        url:
                                          type: string
                                        height:
                                          type: integer
                                        width:
                                          type: integer
                                  name:
                                    type: string
                                  popularity:
                                    type: integer
                                  type:
                                    type: string
                                  uri:
                                    type: string
                            tracks:
                              type: object
                              properties:
                                href:
                                  type: string
                                limit:
                                  type: integer
                                next:
                                  type: string
                                offset:
                                  type: integer
                                previous:
                                  type: string
                                total:
                                  type: integer
                                items:
                                  type: array
                                  items:
                                    type: object
                                    properties:
                                      artists:
                                        type: array
                                        items:
                                          type: object
                                          properties:
                                            external_urls:
                                              type: object
                                              properties:
                                                spotify:
                                                  type: string
                                            href:
                                              type: string
                                            id:
                                              type: string
                                            name:
                                              type: string
                                            type:
                                              type: string
                                            uri:
                                              type: string
                                      available_markets:
                                        type: array
                                        items:
                                          type: string
                                      disc_number:
                                        type: integer
                                      duration_ms:
                                        type: integer
                                      explicit:
                                        type: boolean
                                      external_urls:
                                        type: object
                                        properties:
                                          spotify:
                                            type: string
                                      href:
                                        type: string
                                      id:
                                        type: string
                                      is_playable:
                                        type: boolean
                                      linked_from:
                                        type: object
                                        properties:
                                          external_urls:
                                            type: object
                                            properties:
                                              spotify:
                                                type: string
                                          href:
                                            type: string
                                          id:
                                            type: string
                                          type:
                                            type: string
                                          uri:
                                            type: string
                                      restrictions:
                                        type: object
                                        properties:
                                          reason:
                                            type: string
                                      name:
                                        type: string
                                      preview_url:
                                        type: string
                                      track_number:
                                        type: integer
                                      type:
                                        type: string
                                      uri:
                                        type: string
                                      is_local:
                                        type: boolean

        "401":
            description: unauthorized token
...
  /addLikedSong/{id}/{token}:
    put:
      tags:
        - Liked Songs
      summary: Add a new liked song
      description: Add a new liked song to the user's liked songs
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the song to be added
          schema:
            type: string
        - name: token
          in: path
          required: true
          description: The user's access token
          schema:
            type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
  /listLikedSongs/{offset}/{token}:
    get:
      tags:
        - Liked Songs
      summary: Get a list of the user's liked songs
      description: Get a list of the user's liked songs with optional pagination
      parameters:
        - name: offset
          in: path
          required: true
          description: The index of the first liked song to return
          schema:
            type: integer
            format: int64
        - name: token
          in: path
          required: true
          description: The user's access token
          schema:
            type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
  /deleteLikedSong/{id}/{token}:
    delete:
      tags:
        - Liked Songs
      summary: Delete a liked song
      description: Delete a liked song from the user's liked songs
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the song to be deleted
          schema:
            type: string
        - name: token
          in: path
          required: true
          description: The user's access token
          schema:
            type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
  /isLikedSong/{id}/{token}:
    get:
      tags:
        - Liked Songs
      summary: Check if a song is in the user's liked songs
      description: Check if a song is in the user's liked songs
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the song to be checked
          schema:
            type: string
        - name: token
          in: path
          required: true
          description: The user's access token
          schema:
            type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
...